The answer of DataCamp Exercise
I suggest that it is better to read the answer to the DataCamp Exercise and have a summary of the usage and grammar of the code afterward. 

I have ever searched and figured out the knowledge and finished the exercise. However, I find  it is inefficient.

The notebook  is for self-learning.